+++ .storybook/main.ts
@@ -11,7 +11,9 @@ module.exports = {
       name: '@storybook/addon-docs',
       options: {
         configureJSX: true,
-      },
-    },
+        babelOptions: {},
+        sourceLoaderOptions: null
+      }
+    }
   ],
 };