+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add postbuild script
 - RemoteVideos, RemoteVideo components
 - CSS reset, public domain version
+- Set up customized MDX documentation and add docs for Badge component
 
 ### Changed